fix a warning
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@6363 d7cf8633-e32d-0410-b094-e92efae38249

+++ netconn.c
@@ -2241,7 +2241,7 @@ void NetConn_QueryMasters(qboolean querydp, qboolean queryqw)
 							Con_Printf("Querying master %s (resolved from %s)\n", lookupstring, sv_qwmasters[masternum].string);
 						}
 						masterquerycount++;
-						NetConn_Write(cl_sockets[i], request, strlen(request) + 1, &masteraddress);
+						NetConn_Write(cl_sockets[i], request, (int)strlen(request) + 1, &masteraddress);
 					}
 				}
 			}
